requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain_google_genai
pyPDF2
langchain-community
chroma-db

